Commit for change status

+++ WorkOrderCore/Infrastructure/Persistence/DataContext/JobCards.cs
@@ -11,7 +11,7 @@ namespace WorkOrderCore.Infrastructure.Persistence.DataContext
         }
 
         public int Id { get; set; }
-        public short? BuninessUnitId { get; set; }
+        public int? BuninessUnitId { get; set; }
         public string JobDescription { get; set; }
         public short? JobDuration { get; set; }
         public string JobStatus { get; set; }
